feat: adding buttons, pressure sensors, and OLED SSD1306
I have been working on this commit to optimize and operate the pressure sensor and to control the parameters of that sensor using the OLED and the buttons.
I have created a file that called pages used to control the operation of showing the pages, it's working fine right now it has navigations to pages and lines.
I have create a file that called Butttons and it's used to manage buttons signals and inialize its pins.